app/requirements.txt
Flask
google-cloud-storage
opencv-python
gunicorn